requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.0
pytesseract==0.3.6
requests==2.24.0
scipy==1.5.3
soupsieve==2.0.1
urllib3==1.25.11
Werkzeug==1.0.1
